requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
pillow==10.2.0
pytz==2023.4
sqlparse==0.4.4
tzdata==2023.4
whitenoise==6.6.0
